requirements.txt
Django
zmq
django-adminrestrict
pycryptodome
django-recaptcha
requests
numpy
django-cors-headers 
opencv-python-headless
